+++ Packages/EditorCore/Sources/StringUtils/String+Transform.swift
@@ -1,5 +1,5 @@
 //
-//  String+FullwidthTransform.swift
+//  String+Transform.swift
 //  StringUtils
 //
 //  CotEditor
@@ -9,7 +9,7 @@
 //
 //  ---------------------------------------------------------------------------
 //
-//  © 2014-2024 1024jp
+//  © 2014-2025 1024jp
 //
 //  Licensed under the Apache License, Version 2.0 (the "License");
 //  you may not use this file except in compliance with the License.
@@ -35,6 +35,17 @@ public extension StringProtocol {
         self.unicodeScalars
             .map { $0.convertedToFullwidthRoman(reverse: reverse) ?? $0 }
             .reduce(into: "") { $0.unicodeScalars.append($1) }
+    }
+}
+
+
+public extension String {
+    
+    /// Straightens all curly quotes.
+    var straighteningQuotes: String {
+        
+        self.replacing(/[‘’‚‛]/, with: "'")   // U+2018..201B
+            .replacing(/[“”„‟]/, with: "\"")  // U+201C..201F
     }
 }